requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bugsnag==4.7.1
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.5
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.3
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.41.2
tabulate==0.9.0
typing_extensions==4.12.2
uvicorn==0.32.0
WebOb==1.8.9
yoyo-migrations==9.0.0
zipp==3.21.0
